joomla php8.5 deprecated warnings

+++ libraries/src/Profiler/Profiler.php
@@ -198,7 +198,7 @@ class Profiler
 	/**
 	 * Sets the start time.
 	 *
-	 * @param   double  $startTime  Unix timestamp in microseconds for setting the Profiler start time.
+	 * @param   float  $startTime  Unix timestamp in microseconds for setting the Profiler start time.
 	 * @param   int     $startMem   Memory amount in bytes for setting the Profiler start memory.
 	 *
 	 * @return  $this   For chaining
@@ -207,7 +207,7 @@ class Profiler
 	 */
 	public function setStart($startTime = 0.0, $startMem = 0)
 	{
-		$this->start       = (double) $startTime;
+		$this->start       = (float) $startTime;
 		$this->previousMem = (int) $startMem / 1048576;
 
 		return $this;
